networking_project/requirements.txt
netmiko==4.1.2
pyyaml==6.0
pytest==7.2.0
